// dolphinscheduler-ui/src/service/modules/thirdparty-api-source/types.ts
/*
 * Licensed to the Apache Software Foundation (ASF) under one or more
 * contributor license agreements.  See the NOTICE file distributed with
 * this work for additional information regarding copyright ownership.
 * The ASF licenses this file to You under the Apache License, Version 2.0
 * (the "License"); you may not use this file except in compliance with
 * the License.  You may obtain a copy of the License at
 *
 *    http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

interface ThirdpartyApiSourceReq {
  pageNo: number
  pageSize: number
  searchVal?: string
}

interface ResponseMapping {
  token?: string
  taskInstanceId?: string
  taskInstanceJobId?: string
  type: string
  expected_value: string
}

interface Header {
  [key: string]: string
}

interface Condition {
  path: string
  expected_value: string
}

interface Authentication {
  url: string
  method: string
  queryParams: Record<string, any>
  body?: Record<string, any>
  responseMapping: ResponseMapping
  headers?: Record<string, any>[]
}

interface SubmitTask {
  url: string
  method: string
  queryParams: Record<string, any>
  body?: Record<string, any>
  responseMapping: ResponseMapping
}

interface Polling {
  url: string
  method: string
  queryParams: Record<string, any>
  successConditions: Condition[]
  failureConditions: Condition[]
}

interface Kill {
  url: string
  method: string
  queryParams: Record<string, any>
  body?: Record<string, any>
  responseMapping: ResponseMapping
}

interface ThirdpartyApiSource {
  id?: number
  name: string
  authentication: Authentication
  submit_task: SubmitTask
  polling: Polling
  kill: Kill
  createTime?: string
  updateTime?: string
}

interface ThirdpartyApiSourceRes {
  totalList: ThirdpartyApiSource[]
  total: number
  totalPage: number
  pageSize: number
  currentPage: number
  start: number
}

export {
  ThirdpartyApiSourceReq,
  ThirdpartyApiSource,
  ThirdpartyApiSourceRes,
  Authentication,
  SubmitTask,
  Polling,
  Kill,
  ResponseMapping,
  Header,
  Condition
} 